Final_Project/requirements.txt
pandas 
schedule 
requests
hopsworks
matplotlib.pyplot